refactor: split compression utilities into submodules
